add PackageLink component
