+++ catalog/app/containers/Bucket/PackageTree.js
@@ -1,5 +1,3 @@
-import cx from 'classnames'
-import * as dateFns from 'date-fns'
 import dedent from 'dedent'
 import * as R from 'ramda'
 import * as React from 'react'
@@ -11,7 +9,6 @@ import { Crumb, copyWithoutSpaces, render as renderCrumbs } from 'components/Bre
 import * as Intercom from 'components/Intercom'
 import Placeholder from 'components/Placeholder'
 import * as Preview from 'components/Preview'
-import * as Notifications from 'containers/Notifications'
 import AsyncResult from 'utils/AsyncResult'
 import * as AWS from 'utils/AWS'
 import * as BucketConfig from 'utils/BucketConfig'
@@ -23,8 +20,7 @@ import * as LogicalKeyResolver from 'utils/LogicalKeyResolver'
 import MetaTitle from 'utils/MetaTitle'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import * as PackageUri from 'utils/PackageUri'
-import Link, { linkStyle } from 'utils/StyledLink'
-import copyToClipboard from 'utils/clipboard'
+import Link from 'utils/StyledLink'
 import parseSearch from 'utils/parseSearch'
 import * as s3paths from 'utils/s3paths'
 import usePrevious from 'utils/usePrevious'
@@ -33,6 +29,7 @@ import Code from './Code'
 import CopyButton from './CopyButton'
 import * as FileView from './FileView'
 import Listing from './Listing'
+import PackageLink from './PackageLink'
 import { usePackageUpdateDialog } from './PackageUpdateDialog'
 import PackageCopyDialog from './PackageCopyDialog'
 import Section from './Section'
@@ -41,182 +38,6 @@ import * as errors from './errors'
 import renderPreview from './renderPreview'
 import * as requests from './requests'
 import useViewModes from './viewModes'
-
-function useRevisionsData({ bucket, name }) {
-  const req = AWS.APIGateway.use()
-  return useData(requests.getPackageRevisions, { req, bucket, name, perPage: 5 })
-}
-
-const useRevisionInfoStyles = M.makeStyles((t) => ({
-  revision: {
-    ...linkStyle,
-    alignItems: 'center',
-    display: 'inline-flex',
-  },
-  mono: {
-    fontFamily: t.typography.monospace.fontFamily,
-  },
-  line: {
-    whiteSpace: 'nowrap',
-  },
-  secondaryText: {
-    display: 'block',
-    height: 40,
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-  },
-  list: {
-    width: 420,
-  },
-}))
-
-function RevisionInfo({ revisionData, revision, bucket, name, path }) {
-  const { urls } = NamedRoutes.use()
-  const { push } = Notifications.use()
-  const classes = useRevisionInfoStyles()
-
-  const listRef = React.useRef()
-  const [anchor, setAnchor] = React.useState()
-  const [opened, setOpened] = React.useState(false)
-  const open = React.useCallback(() => setOpened(true), [])
-  const close = React.useCallback(() => setOpened(false), [])
-
-  const revisionsData = useRevisionsData({ bucket, name })
-  const data = revisionsData.case({
-    Ok: (revisions) =>
-      revisionData.case({
-        Ok: ({ hash }) =>
-          AsyncResult.Ok(revisions.map((r) => ({ ...r, selected: r.hash === hash }))),
-        Err: () => AsyncResult.Ok(revisions),
-        _: R.identity,
-      }),
-    _: R.identity,
-  })
-
-  const getHttpsUri = (r) =>
-    `${window.origin}${urls.bucketPackageTree(bucket, name, r.hash, path)}`
-
-  const copyHttpsUri = (r, containerRef) => (e) => {
-    e.preventDefault()
-    copyToClipboard(getHttpsUri(r), { container: containerRef && containerRef.current })
-    push('Canonical URI copied to clipboard')
-  }
-
-  return (
-    <>
-      {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions */}
-      <span
-        className={classes.revision}
-        onClick={open}
-        ref={setAnchor}
-        title={revision.length > 10 ? revision : undefined}
-      >
-        {R.take(10, revision)} <M.Icon>expand_more</M.Icon>
-      </span>
-
-      {revisionData.case({
-        Ok: (r) => (
-          <M.IconButton
-            size="small"
-            title="Copy package revision's canonical catalog URI to the clipboard"
-            href={getHttpsUri(r)}
-            onClick={copyHttpsUri(r)}
-            style={{ marginTop: -4, marginBottom: -4 }}
-          >
-            <M.Icon>link</M.Icon>
-          </M.IconButton>
-        ),
-        _: () => null,
-      })}
-
-      <M.Popover
-        open={opened && !!anchor}
-        anchorEl={anchor}
-        onClose={close}
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-        transformOrigin={{ vertical: 'top', horizontal: 'center' }}
-      >
-        <M.List className={classes.list} ref={listRef}>
-          {AsyncResult.case(
-            {
-              Ok: R.ifElse(
-                R.isEmpty,
-                () => (
-                  <M.ListItem>
-                    <M.ListItemText
-                      primary="No revisions found"
-                      secondary="Looks like this package has been deleted"
-                    />
-                  </M.ListItem>
-                ),
-                R.map((r) => (
-                  <M.ListItem
-                    key={r.hash}
-                    button
-                    onClick={close}
-                    selected={r.selected}
-                    component={RRLink}
-                    to={urls.bucketPackageTree(bucket, name, r.hash, path)}
-                  >
-                    <M.ListItemText
-                      primary={dateFns.format(r.modified, 'MMMM do yyyy - h:mma')}
-                      secondary={
-                        <span className={classes.secondaryText}>
-                          <span className={classes.line}>
-                            {r.message || <i>No message</i>}
-                          </span>
-                          <br />
-                          <span className={cx(classes.line, classes.mono)}>{r.hash}</span>
-                        </span>
-                      }
-                    />
-                    <M.ListItemSecondaryAction>
-                      <M.IconButton
-                        title="Copy package revision's canonical catalog URI to the clipboard"
-                        href={getHttpsUri(r)}
-                        onClick={copyHttpsUri(r, listRef)}
-                      >
-                        <M.Icon>link</M.Icon>
-                      </M.IconButton>
-                    </M.ListItemSecondaryAction>
-                  </M.ListItem>
-                )),
-              ),
-              Err: () => (
-                <M.ListItem>
-                  <M.ListItemIcon>
-                    <M.Icon>error</M.Icon>
-                  </M.ListItemIcon>
-                  <M.Typography variant="body1">Error fetching revisions</M.Typography>
-                </M.ListItem>
-              ),
-              _: () => (
-                <M.ListItem>
-                  <M.ListItemIcon>
-                    <M.CircularProgress size={24} />
-                  </M.ListItemIcon>
-                  <M.Typography variant="body1">Fetching revisions</M.Typography>
-                </M.ListItem>
-              ),
-            },
-            data,
-          )}
-          <M.Divider />
-          <M.ListItem
-            button
-            onClick={close}
-            component={RRLink}
-            to={urls.bucketPackageRevisions(bucket, name)}
-          >
-            <M.Box textAlign="center" width="100%">
-              Show all revisions
-            </M.Box>
-          </M.ListItem>
-        </M.List>
-      </M.Popover>
-    </>
-  )
-}
 
 function ExposeLinkedData({ bucketCfg, bucket, name, hash, modified }) {
   const sign = AWS.Signer.useS3Signer()
@@ -654,9 +475,6 @@ function ResolverProvider({ bucket, name, hash, children }) {
 }
 
 const useStyles = M.makeStyles({
-  name: {
-    wordBreak: 'break-all',
-  },
   alertMsg: {
     overflow: 'hidden',
     textOverflow: 'ellipsis',
@@ -757,17 +575,7 @@ function PackageTree({ bucket, mode, name, revision, path, resolvedFrom }) {
           </Lab.Alert>
         </M.Box>
       )}
-      <M.Typography variant="body1">
-        <Link to={urls.bucketPackageDetail(bucket, name)} className={classes.name}>
-          {name}
-        </Link>
-        {' @ '}
-        <RevisionInfo
-          {...{ revisionData, revision, bucket, name, path }}
-          key={`revinfo:${revisionListKey}`}
-        />
-      </M.Typography>
-
+      <PackageLink {...{ bucket, name, path, revision, revisionListKey, revisionData }} />
       {revisionData.case({
         Ok: ({ hash }) => (
           <ResolverProvider {...{ bucket, name, hash }}>
